move loaders into subdir
