+++ ml_datasets/__init__.py
@@ -1,10 +1,2 @@
 from ._registry import register_loader as loaders
-from .imdb import imdb
-from .mnist import mnist
-from .quora import quora_questions
-from .reuters import reuters
-from .snli import snli
-from .stack_exchange import stack_exchange
-from .universal_dependencies import ud_ancora_pos_tags, ud_ewtb_pos_tags
-from .dbpedia import dbpedia
-from .cmu import cmu
+from ml_datasets.loaders.stack_exchange import stack_exchange
